+++ mp/src/public/vgui/ISurface.h
@@ -41,9 +41,9 @@ class Image;
 class Point;
 
 // handles
-typedef unsigned long HCursor;
-typedef unsigned long HTexture;
-typedef unsigned long HFont;
+typedef uint32_t HCursor;
+typedef uint32_t HTexture;
+typedef uint32_t HFont;
 
 
 //SRC only defines
