+++ mp/src/public/vgui/VGUI.h
@@ -50,17 +50,17 @@ namespace vgui
 {
 // handle to an internal vgui panel
 // this is the only handle to a panel that is valid across dll boundaries
-typedef unsigned int VPANEL;
+typedef uintptr_t VPANEL;
 
 // handles to vgui objects
 // NULL values signify an invalid value
-typedef unsigned long HScheme;
+typedef uint32_t HScheme;
 // Both -1 and 0 are used for invalid textures. Be careful.
-typedef unsigned long HTexture;
-typedef unsigned long HCursor;
-typedef unsigned long HPanel;
+typedef uint32_t HTexture;
+typedef uint32_t HCursor;
+typedef uint32_t HPanel;
 const HPanel INVALID_PANEL = 0xffffffff;
-typedef unsigned long HFont;
+typedef uint32_t HFont;
 const HFont INVALID_FONT = 0; // the value of an invalid font handle
 }
